+++ oh-my-zsh/custom/lazygit.zsh
@@ -1,0 +1,6 @@
+#!/bin/zsh
+
+export XDG_CONFIG_HOME="$HOME/.config"
+
+# Start lazygit with 'lg'
+alias lg='lazygit'